Modified the adding a book function to see that if it there is the exact same name and author, do not add it. Throws a dupe error and frees the allocated memory that was given for that new entry.
Also modified the removeUser function to see if one or more options at that index are available as many users can have the same last name. IF there are more than 1, it will ask the user to rpecisely choose which to delete and frees the memory. If there is only one, it auto deletes it.